refactor: reorganize project configuration and remove unused files
- Update .pre-commit-config.yaml to use official mypy mirror and add type stub dependencies
- Remove local mypy hook and replace with upstream configuration
- Update Makefile to simplify Dagster development commands
- Remove unused sqlmesh_types.py and workspace.yaml files
- Migrate SQLMeshModelDep and SQLMeshMultiAssetOptions to types.py
- Update import paths in dagster.py to reflect file reorganization
