--- sample/workspace.yaml
@@ -1,4 +1,0 @@
-load_from:
-  - python_module:
-      module_name: dagster_project.definitions
-      working_directory: .
